backend/pyproject.toml
[project]
name = "cad-backend"
version = "0.1.0"
description = "Backend service for the CAD project"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "Pillow",
    "python-multipart",
    "uvicorn",
]

[project.optional-dependencies]
test = [
    "httpx",
    "pytest",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
